+++ resolwe/flow/elastic_indexes/base.py
@@ -6,6 +6,14 @@ from resolwe.elastic.fields import Name, Slug, User
 from resolwe.permissions.shortcuts import get_users_with_permission
 from resolwe.permissions.utils import get_full_perm
 
+# pylint: disable=invalid-name
+sorting_analyzer = dsl.analyzer(
+    'sorting_analyzer',
+    tokenizer='keyword',
+    filter=['lowercase', 'trim'],
+)
+# pylint: enable=invalid-name
+
 
 class BaseDocument(indices.BaseDocument):
     """Base search document."""
@@ -20,7 +28,7 @@ class BaseDocument(indices.BaseDocument):
     contributor_name = User()
     # We use a separate field for contributor sorting because we use an entirely
     # different value for it (the display name).
-    contributor_sort = dsl.Keyword()
+    contributor_sort = dsl.Text(analyzer=sorting_analyzer)
     owner_ids = dsl.Integer(multi=True)
     owner_names = User(multi=True)
 
@@ -32,8 +40,8 @@ class BaseIndexMixin(object):
         """Generate display name for contributor."""
         user = obj.contributor
 
-        if user.first_name:
-            return '{} {}'.format(user.first_name, user.last_name)
+        if user.first_name or user.last_name:
+            return user.get_full_name()
         else:
             return user.username